+++ mybatis-r2dbc/src/main/java/pro/chenggang/project/reactive/mybatis/support/r2dbc/executor/key/NoKeyR2dbcKeyGenerator.java
@@ -36,8 +36,8 @@ public class NoKeyR2dbcKeyGenerator implements R2dbcKeyGenerator {
     }
 
     @Override
-    public Integer processGeneratedKeyResult(RowResultWrapper rowResultWrapper, Object parameter) {
-        return 0;
+    public Long processGeneratedKeyResult(RowResultWrapper rowResultWrapper, Object parameter) {
+        return 0L;
     }
 
     private static class InstanceHolder {
