+++ mybatis-r2dbc/src/main/java/pro/chenggang/project/reactive/mybatis/support/r2dbc/executor/key/R2dbcKeyGenerator.java
@@ -37,5 +37,5 @@ public interface R2dbcKeyGenerator {
      * @param parameter        the parameter
      * @return the mono
      */
-    Integer processGeneratedKeyResult(RowResultWrapper rowResultWrapper, Object parameter);
+    Long processGeneratedKeyResult(RowResultWrapper rowResultWrapper, Object parameter);
 }
