+++ mybatis-r2dbc/src/main/java/pro/chenggang/project/reactive/mybatis/support/r2dbc/executor/support/R2dbcStatementLog.java
@@ -56,7 +56,7 @@ public class R2dbcStatementLog {
      *
      * @param updateCount the update count
      */
-    public void logUpdates(Integer updateCount) {
+    public void logUpdates(Long updateCount) {
         debug("   Updates: " + updateCount, false);
     }
